refactor(file): structure
