+++ src/interfaces/comment.ts
@@ -1,11 +1,11 @@
-import { Event } from "./event";
+import { Event } from "./event/event";
 import { User } from "./user";
 
 export interface Comment {
     id: number;
     content: string;
-    event?: Event;
-    user?: User;
+    event: Event;
+    user: User;
     parent?: Comment;
     replies: Comment[];
 };
