+++ src/interfaces/event/event-tag.ts
@@ -1,5 +1,5 @@
 import { Event } from "./event";
-import { Tag } from "./tag";
+import { Tag } from "../tag/tag";
 
 export interface EventTag {
     id: number;
