+++ src/interfaces/tag/tag-follow.ts
@@ -1,5 +1,5 @@
 import { Tag } from "./tag";
-import { User } from "./user";
+import { User } from "../user";
 
 export interface TagFollow {
     id: number;
